# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Health Plan Factory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans/generate:
    post:
      operationId: generatePlan
      tags: [plans]
      summary: Generate and persist a wellness plan from an intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePlanBody"
      responses:
        "201":
          description: Plan generated and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithItems"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a saved plan with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithItems"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update plan status (save/activate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanRecord"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: List member's progress logs
      parameters:
        - name: profileId
          in: query
          required: true
          schema:
            type: string
        - name: planId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Progress logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressLogRecord"
    post:
      operationId: createProgressLog
      tags: [progress]
      summary: Log a wellness session or check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgressLogBody"
      responses:
        "201":
          description: Progress logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressLogRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
...
    GeneratePlanBody:
      type: object
      properties:
        intakeId:
          type: string
        profileId:
          type: string
        budget:
          type: integer
        goals:
          type: array
          items:
            type: string
        conditions:
          type: array
          items:
            type: string
        preferences:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        zipCode:
          type: string
        radius:
          type: integer
        telehealth:
          type: boolean
      required:
        - budget
        - goals
        - conditions
        - preferences
        - exclusions
        - telehealth
    PlanRecord:
      type: object
      properties:
        id:
          type: string
        profileId:
          type: string
          nullable: true
        intakeId:
          type: string
          nullable: true
        status:
          type: string
        totalMonthlyCost:
          type: integer
        budgetUtilization:
          type: integer
        budget:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - totalMonthlyCost
        - budgetUtilization
        - budget
        - createdAt
        - updatedAt
    PlanItemRecord:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        modalityId:
          type: string
        score:
          type: integer
        frequency:
          type: string
        estimatedMonthlyCost:
          type: integer
        rationale:
          type: string
        isDeprioritized:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - planId
        - modalityId
        - score
        - frequency
        - estimatedMonthlyCost
        - rationale
        - isDeprioritized
        - sortOrder
    PlanWithItems:
      type: object
      properties:
        plan:
          $ref: "#/components/schemas/PlanRecord"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlanItemRecord"
      required:
        - plan
        - items
    UpdatePlanBody:
      type: object
      properties:
        status:
          type: string
...
    ProgressLogRecord:
      type: object
      properties:
        id:
          type: string
        profileId:
          type: string
        planId:
          type: string
          nullable: true
        modalityId:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
        sessionDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - profileId
        - createdAt
    CreateProgressLogBody:
      type: object
      properties:
        profileId:
          type: string
        planId:
          type: string
        modalityId:
          type: string
        note:
          type: string
        rating:
          type: integer
        mood:
          type: integer
          minimum: 1
          maximum: 10
          description: Self-reported mood score (1-10)
        pain:
          type: integer
          minimum: 1
          maximum: 10
          description: Self-reported pain level (1-10)
        energy:
          type: integer
          minimum: 1
          maximum: 10
          description: Self-reported energy level (1-10)
        sessionDate:
          type: string
          format: date-time
        sessionCostCents:
          type: integer
          minimum: 0
          description: >
            Actual out-of-pocket session cost in cents paid by the member.
            When provided, this amount is deducted from the member's employer
            wellness stipend balance (if enrolled). Do not use modality cost
            estimates for stipend deduction — always supply the real session cost.
      required:
        - profileId
...